Develop

+++ src/main/java/com/nnk/springboot/repositories/RatingRepository.java
@@ -3,6 +3,12 @@ package com.nnk.springboot.repositories;
 import com.nnk.springboot.domain.Rating;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RatingRepository extends JpaRepository<Rating, Integer> {
+/**
+ * Repository interface for Rating entity.
+ * 
+ * Provides CRUD operations and database access for Rating entities.
+ * Extends JpaRepository to inherit standard Spring Data JPA functionality.
+ */
+public interface RatingRepository extends JpaRepository<Rating, Long> {
 
 }
